Static building for Netlify and prepation for serverless
There is now a very simple build script to take the current Express index template and render it to HTML. This will be used to to demonstrate the Netlify Doppler intergration to sync secrets to Netlify in order to be used by the build process.

In preparation for being able to eventually run the Express app on Netlify functions and Lambda, the app needs to be able to run as a standard Express app, as well as using custom Netlify/Serverless wrappers.

That's why the server is now run with `node ./src/server.js`.

Also moved the Heroku related portion of the Makefile into its own folder. It needs to be updated now that Heroku sync is only one way, but that's a task for another PR.