Added npm script for installing the Doppler CLI standalone
This is for restirctive Linux cloud environments where the Doppler CLI can't be installed via the usual means.

It puts the doppler binary at the root of the repo directory so it can be accessed during the build and server runtime if they happen to occur on different machines. For example, Render runs the build process on one machine, zips up the working directory, then deploys that to a different server.